requirements.txt
vk_api==11.9.9
psycopg2-binary==2.9.5
bs4